feat(coordinator,cli,miner): P2.2 TEE attestation for confidential jobs
- Constraints gain `tee_attestation_required` and `tee_enclave_id`.
- JobView/JobResultSubmit carry `tee_status`, `tee_attestation_id` and `tee_quote`.
- Miner result path generates/stores/verifies a TEE attestation when required.
- Escrow release is blocked unless `tee_status: verified` for confidential jobs.
- `production_miner.py` auto-generates a simulated TEE quote for TEE jobs.
- `aitbc ai submit` gains `--tee-attestation-required` and `--tee-enclave-id`.

+++ apps/exchange/simple_exchange/handlers/bridge.py
@@ -268,9 +268,9 @@ class BridgeMixin:
 
             rates: dict[str, float] = {}
             if eth_usd and ait_usd and ait_usd.price > 0:
-                rates["ETH::AITBC"] = round(eth_usd.price / ait_usd.price, 8)
+                rates["ETH::AITBC"] = float(round(eth_usd.price / ait_usd.price, 8))
             if eth_usd and ait_usd and eth_usd.price > 0:
-                rates["AITBC::ETH"] = round(ait_usd.price / eth_usd.price, 8)
+                rates["AITBC::ETH"] = float(round(ait_usd.price / eth_usd.price, 8))
 
             self.send_json_response({
                 "rates": rates,
